requirements.txt
altgraph==0.17
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
future==0.18.2
idna==2.10
importlib-metadata==4.5.0
iso8601==0.1.14
Jinja2==3.0.1
MarkupSafe==2.0.1
pip-search==0.0.7
pkg-resources==0.0.0
Pygments==2.9.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pyserial==3.5
PyYAML==5.4.1
requests==2.25.1
rich==10.3.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
